% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_gradient.R
\name{calc_gradient}
\alias{calc_gradient}
\title{Calculate the gradient}
\usage{
calc_gradient(y, f_guess, inv_cov, mu)
}
\arguments{
\item{y}{vector, count for a specific user and time}

\item{f_guess}{vector, prior estimate of the mode of the latent GP}

\item{inv_cov}{matrix, current prior for GP precision matrix}

\item{mu}{vector, linear predictor without the GP}
}
\value{
vector
}
\description{
Calculate the gradient
}
\examples{
inv_cov <- qr.solve(matrix(c(3, 1, 1, 3), ncol = 2))
calc_gradient(c(1, 2), c(2, 3), cov_m, c(-5, -4))
}
