% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newton_raphson_functions.R
\name{calc_hessian}
\alias{calc_hessian}
\title{Caluate the hessian}
\usage{
calc_hessian(f_cur, inv_cov, mu)
}
\arguments{
\item{f_cur}{vector, current estimate of the mode of the latent GP}

\item{inv_cov}{matrix, prior for GP precision matrix}

\item{mu}{vector, linear predictor without the GP}
}
\value{
matrix
}
\description{
Caluate the hessian
}
\examples{
inv_cov <- qr.solve(matrix(c(3, 1, 1, 3), ncol = 2))
hessian(c(2, 3), inv_cov, c(-5, -4))
}
