% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{diag_each_chain}
\alias{diag_each_chain}
\title{Attain diagnostics for each chain}
\usage{
diag_each_chain(draws, accepts, was_burn_in_kept, burn_in)
}
\arguments{
\item{draws}{matrix of pmmh draws}

\item{accepts}{matrix of acceptances}

\item{was_burn_in_kept}{logical, were burn in values}

\item{burn_in}{burn in}
}
\value{
list
}
\description{
Attain diagnostics for each chain
}
