% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{evaluate_pmmh}
\alias{evaluate_pmmh}
\title{Build a dataset for easy testing}
\usage{
evaluate_pmmh(pmmh_res, gen_data_obj, burn_in = 51)
}
\arguments{
\item{pmmh_res}{list of results from the pmmh_function}

\item{burn_in}{scalar, number of draws to discard}

\item{data}{scalar}

\item{mean_vec}{vector, linear predictor without the GP}
}
\value{
list with data matrix vector or true values
}
\description{
Build a dataset for easy testing
}
