% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{gen_data}
\alias{gen_data}
\title{Build a dataset for easy testing}
\usage{
gen_data(
  sigma,
  phi,
  patients,
  time_points_per_patient,
  length_gp = 10,
  seed = 28,
  alpha_hier,
  beta_hier,
  offset_vec = NULL,
  num_add_covar = NULL
)
}
\arguments{
\item{sigma}{scalar}

\item{phi}{scalar}

\item{data_obj}{list of objects to make relevant data matrices}

\item{mean_vec}{vector, linear predictor without the GP}
}
\value{
list with data matrix vector or true values
}
\description{
Build a dataset for easy testing
}
