% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pmll_estim.R
\name{get_pmll_estim}
\alias{get_pmll_estim}
\title{Get estimates of the marginal likelihood values}
\usage{
get_pmll_estim(y_matrix, cov_mat, inv_cov_mat, mean_mat, nsim)
}
\arguments{
\item{y_matrix}{outcome matrix where each column is a multivariate observation}

\item{cov_mat}{latent GP covariance matrix}

\item{inv_cov_mat}{latent GP precision matrix}

\item{mean_mat}{linear predictors ommitting latent GP}

\item{nsim}{scalar, number of importance samples}
}
\value{
vector
}
\description{
Get estimates of the marginal likelihood values
}
\examples{
get_pmll_estim(y_mat, cov_mat, inv_cov_mat, mean_mat, nsim = 20)
}
