% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_post_estim.R
\name{get_post_estim}
\alias{get_post_estim}
\title{Get values for MH accept step}
\usage{
get_post_estim(data, param_vec, param_name, subj_index, nsim, pmll_vec = NULL)
}
\arguments{
\item{data}{list of objects}

\item{param_vec}{vector of proposals}

\item{param_name}{char, name of parameter proposed}

\item{subj_index}{scalar, index of subject}

\item{nsim}{scalar, number of importance samples}

\item{pmll_vec}{vector of previous pmll estimates}
}
\value{
vector
}
\description{
Get values for MH accept step
}
