% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_post_estimate.R
\name{get_post_estimate}
\alias{get_post_estimate}
\title{Get values for MH accept step}
\usage{
get_post_estimate(
  data,
  param_vec,
  param_name,
  pat_index,
  patients,
  nsim,
  ll_marg_vec = NULL
)
}
\arguments{
\item{nsim}{scalar, number of importance samples}

\item{gen_data_mat_obj}{object with list of relevant matrices}
}
\value{
vector
}
\description{
Get values for MH accept step
}
\examples{
get_post_estimate(data, param_vec, param_name, param_)
}
