% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newton_raphson_functions.R
\name{get_posterior_modes}
\alias{get_posterior_modes}
\title{Get posterior modes using laplace approximation}
\usage{
get_posterior_modes(y_mat, inv_cov_mat, mean_mat)
}
\arguments{
\item{y_mat}{matrix of count values where columns are the multivariate observations}

\item{mean_mat}{matrix of linear predictors without the latent GP values}

\item{cov_mat}{prior precision matrix for the latent GP}
}
\value{
vector
}
\description{
Get posterior modes using laplace approximation
}
\examples{
y <- matrix(1:4, ncol = 2)
inv_cov <- qr.solve(matrix(c(1, 0.5, 0.5, 1), ncol = 2))
mean_mat <- y - 0.5
get_posterior_modes(y, inv_cov, mean_mat)
}
