% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxim_test.R
\name{maxim_test}
\alias{maxim_test}
\title{Get values for MH accept step}
\usage{
maxim_test(dat, param_vec, num_subjects, nsim = 50)
}
\arguments{
\item{dat}{list of objects}

\item{param_vec}{vector of proposals}
}
\value{
vector
}
\description{
Get values for MH accept step
}
