% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmmh.R
\name{pmmh}
\alias{pmmh}
\title{Get values for MH accept step}
\usage{
pmmh(
  data,
  distance_mat,
  outcome = "y",
  subject_index = "patient",
  time_index = "time",
  position_index = "position",
  ndraws = 200,
  nsim = 10,
  offset_term = NULL
)
}
\arguments{
\item{data}{data.frame or matrix with described column names}

\item{distance_mat}{matrix, distance between multivariate outcomes}

\item{outcome}{char, name of the outcome variable}

\item{subject_index}{char, name of the column denoting subjects}

\item{time_index}{char, name of the column denoting time}

\item{position_index}{char, name of the column denoting position}

\item{ndraws}{number of draws}

\item{nsim}{number of importance samples}

\item{offset_term}{char, name of column that is used as an offset}
}
\value{
list
}
\description{
Get values for MH accept step
}
