% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{pmmh_diags}
\alias{pmmh_diags}
\title{Attain diagnostics for all chaings}
\usage{
pmmh_diags(pmmh_list)
}
\arguments{
\item{pmmh_list}{list returned from pmmh}
}
\value{
list
}
\description{
Attain diagnostics for all chaings
}
