% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmmh_single_chain.R
\name{pmmh_single_chain}
\alias{pmmh_single_chain}
\title{Run pmmh for a single chain}
\usage{
pmmh_single_chain(single_chain_list)
}
\arguments{
\item{single_chain_list}{list passed from pmmh}
}
\value{
list
}
\description{
Run pmmh for a single chain
}
