% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{prepare_data}
\alias{prepare_data}
\title{Prepare data for loop within pmmh}
\usage{
prepare_data(
  data,
  distance_mat,
  length_gp,
  num_subjects,
  outcome,
  subject_index,
  time_index,
  position_index,
  offset_term
)
}
\arguments{
\item{data}{list of objects to make relevant data matrices}

\item{distance_mat}{matrix, distance between multivariate outcomes}

\item{length_gp}{scalar, length of the multivariate outcomes}

\item{num_subjects}{scalar, number of subjects}
}
\value{
list of matrices
}
\description{
Prepare data for loop within pmmh
}
