% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{prepare_storage}
\alias{prepare_storage}
\title{Prepare storage for results}
\usage{
prepare_storage(data, ndraws, num_subjects, nsim, burn_in, keep_burn_in)
}
\arguments{
\item{data}{list of objects to make relevant storage}

\item{ndraws}{scalar, number of mh iterations}

\item{num_subjects}{scalar, number of subjects}

\item{nsim}{scalar, number of importance samples}

\item{burn_in}{scalar, number of draws to discard}

\item{keep_burn_in}{logical, whether burn_in draws should be kept}
}
\value{
list for consumption by pmmh_single_chain
}
\description{
Prepare storage for results
}
