% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newton_raphson_functions.R
\name{single_step}
\alias{single_step}
\title{Take a single step in the newton raphson algorithm}
\usage{
single_step(y, f_cur, inv_cov, mu, step_halve = FALSE)
}
\arguments{
\item{y}{vector, count for a specific user and time}

\item{f_cur}{vector, current estimate of the mode of the latent GP}

\item{inv_cov}{matrix, prior for GP precision matrix}

\item{mu}{vector, linear predictor without the GP}

\item{step_halve}{logical, step halving}
}
\value{
vector
}
\description{
Take a single step in the newton raphson algorithm
}
\examples{
inv_cov <- qr.solve(matrix(c(3, 1, 1, 3), ncol = 2))
single_step(c(1, 2), c(2, 3), inv_cov, c(-5, -4))
}
