% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pmll_estim.R
\name{get_pmll_estim}
\alias{get_pmll_estim}
\title{Get estimates of the marginal likelihood values}
\usage{
get_pmll_estim(get_data_mat_obj, nsim = 500)
}
\arguments{
\item{nsim}{scalar, number of importance samples}

\item{gen_data_mat_obj}{object with list of relevant matrices}
}
\value{
vector
}
\description{
Get estimates of the marginal likelihood values
}
\examples{
#' get_pmll_estim(data_object, nsim = 20)
}
