% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmmh.R
\name{pmmh}
\alias{pmmh}
\title{Get values for MH accept step}
\usage{
pmmh(gen_data_obj, ndraws = 300, nsim = 25)
}
\arguments{
\item{gen_data_obj}{list of objects}

\item{ndraws}{number of draws}

\item{nsim}{number of importance samples}
}
\value{
list
}
\description{
Get values for MH accept step
}
